requirements.txt
pywin32
psutil
discord.py